fix use bookmark
