+++ web/client-v2/src/features/bookmarks/components/Dialog/DeleteBookmarkAlertDialog/DeleteBookmarkAlertDialog.tsx
@@ -3,7 +3,7 @@ import { usePathname } from "next/navigation";
 import { FC } from "react";
 import { FcBookmark } from "react-icons/fc";
 
-import { useBookmark } from "@/features/bookmarks/hooks/useBookmark";
+import { useBookmarkMutation } from "@/features/bookmarks/hooks/useBookmarkMutation";
 
 import {
   AlertDialog,
@@ -35,7 +35,7 @@ export const DeleteBookmarkAlertDialog: FC<DeleteBookmarkAlertDialogProps> = ({
   bookmarkTitle,
 }) => {
   const pathname = usePathname();
-  const { handleRemoveBookmark } = useBookmark(pathname);
+  const { handleRemoveBookmark } = useBookmarkMutation();
   return (
     <AlertDialog>
       <TooltipProvider>
@@ -59,13 +59,15 @@ export const DeleteBookmarkAlertDialog: FC<DeleteBookmarkAlertDialogProps> = ({
             This action cannot be undone. This will permanently remove your data
             from our servers.
             <br></br>
-            Delete bookmark title is{" "}
-            <span className="font-bold text-white">{`「${bookmarkTitle}」`}</span>
+            The title of the bookmark to be deleted is{" "}
+            <span className="font-bold text-white">{`'${bookmarkTitle}'`}</span>
           </AlertDialogDescription>
         </AlertDialogHeader>
         <AlertDialogFooter>
           <AlertDialogCancel>CANCEL</AlertDialogCancel>
-          <AlertDialogAction onClick={() => handleRemoveBookmark(bookmarkId)}>
+          <AlertDialogAction
+            onClick={() => handleRemoveBookmark(bookmarkId, bookmarkTitle)}
+          >
             DELETE
           </AlertDialogAction>
         </AlertDialogFooter>
